feat: refactor for tools
