+++ eelbrain/_trf/_code.py
@@ -6,7 +6,7 @@ import numpy as np
 
 from .._data_obj import Dataset
 from .._experiment.definitions import CodeBase, CodeError
-from ._ndvar import SHUFFLE_METHODS as NDVAR_SHUFFLE_METHODS
+from .tools.ndvar import SHUFFLE_METHODS as NDVAR_SHUFFLE_METHODS
 
 
 VALUE_SHUFFLE_METHODS = ('permute', 'remask', 'relocate')
