+++ eelbrain/_trf/tools/__init__.py
@@ -1,0 +1,4 @@
+"""TRF-local helper modules."""
+
+from .ndvar import SHUFFLE_METHODS, shuffle
+from .numpy_funcs import arctanh
