+++ eelbrain/_trf/tools/ndvar.py
